requirements.txt
Flask
python-dotenv
google-genai
pillow
protobuf
requests
google-generativeai